UInt32 everywhere, add test

+++ src/Storages/MergeTree/MergeTreeSettings.cpp
@@ -71,7 +71,7 @@ namespace ErrorCodes
     Minimum number of bytes/rows in a data part that can be stored in `Wide`
     format. You can set one, both or none of these settings.
     )", 0) \
-    DECLARE(UInt64, min_level_for_wide_part, 0, R"(
+    DECLARE(UInt32, min_level_for_wide_part, 0, R"(
     Minimal part level to create a data part in `Wide` format instead of `Compact`.
     )", 0) \
     DECLARE(UInt64, min_rows_for_wide_part, 0, R"(
@@ -208,7 +208,7 @@ namespace ErrorCodes
     Only available in ClickHouse Cloud. Minimal uncompressed size in bytes to
     use full type of storage for data part instead of packed
     )", 0) \
-    DECLARE(UInt64, min_level_for_full_part_storage, 0, R"(
+    DECLARE(UInt32, min_level_for_full_part_storage, 0, R"(
     Only available in ClickHouse Cloud. Minimal part level to
     use full type of storage for data part instead of packed
     )", 0) \
